djangoapp/requirements.txt
asgiref==3.8.1
channels==4.1.0
daphne>=4.0.0
Django==5.1.1
sqlparse==0.5.1
tzdata==2024.1

psycopg2-binary>=2.9.6,<2.10 #for postgre database
dnspython #for verify real email
Pillow #for imagens in database

requests
requests-oauthlib
ngrok
whitenoise #pra que mesmo? 

autobahn>=22.3 #for debug sql
django-extensions>=3.2.3,<3.3 #for debug sql
setuptools>=65.5.0,<65.6 #for debug sql
MarkupSafe>=3.0.1,<3.1 #for debug sql
typing_extensions>=4.12.2,<4.13 #for debug sql
Werkzeug>=3.0.4,<3.1 #Werkzeug is a comprehensive WSGI web application library. 


djangorestframework #permite criar APIs RESTful no Django de forma fácil e estruturada.

